Add files via upload
At the time this was a very difficult program to start because I had to figure out how to loop through two parallel arrays. I ended up putting all the features in thee same loop and recycled the iteration variables. It was a clean and simple solution but I understand the limitations of this coding style.